zero: ported to net standard 2.1

+++ zero.cocoon/models/CcDiscoveries.cs
@@ -284,7 +284,7 @@ namespace zero.cocoon.models
                     }
                     finally
                     {
-                        BufferOffset += (uint)read;
+                        BufferOffset += read;
                     }
 
                     if (read == 0)
